Added interim detail view.
