+++ main/templates/main/detail.html
@@ -7,7 +7,7 @@
 
 
 <dl class="govuk-summary-list">
-    {% for key, value in object.get_detail_list %}
+    {% for key, value in display_list %}
 <div class="govuk-summary-list__row">
     <dt class="govuk-summary-list__key">
         {{ key }}
